Moved class SIMFractureQstatic to a separate header file
